After code review:  - extract update as template method  - use indexOf() in delete and get
