+++ src/ru/javawebinar/basejava/storage/ArrayStorage.java
@@ -2,22 +2,18 @@ package ru.javawebinar.basejava.storage;
 
 import ru.javawebinar.basejava.model.Resume;
 
-import java.util.Arrays;
-
 /**
  * Array based storage for Resumes
  */
 public class ArrayStorage extends AbstractArrayStorage {
 
-
     @Override
-    protected void __save(int index, Resume r) {
+    protected void saveIntoArray(int index, Resume r) {
         storage[size] = r;
     }
 
 
-    @Override
-    protected void __delete(int index) {
+    protected void deleteArray(int index) {
         storage[index] = storage[size - 1];
     }
