+++ src/ru/javawebinar/basejava/storage/SortedArrayStorage.java
@@ -6,16 +6,14 @@ import java.util.Arrays;
 
 public class SortedArrayStorage extends AbstractArrayStorage {
 
-
     @Override
-    protected void __save(int index, Resume r) {
+    protected void saveIntoArray(int index, Resume r) {
         int newIndex = -index - 1;
         System.arraycopy(storage, newIndex, storage, newIndex + 1, size - newIndex);
         storage[newIndex] = r;
     }
 
-    @Override
-    protected void __delete(int index) {
+    protected void deleteArray(int index) {
         System.arraycopy(storage, index + 1, storage, index, size - index);
     }
