+++ src/ru/javawebinar/basejava/storage/Storage.java
@@ -18,5 +18,4 @@ public interface Storage {
 
     int size();
 
-    int capacity();
 }
