+++ test/ru/javawebinar/basejava/storage/ListStorageTest.java
@@ -1,6 +1,6 @@
 package ru.javawebinar.basejava.storage;
 
-public class ListStorageTest extends AbstractArrayStorageTest {
+public class ListStorageTest extends AbstractStorageTest {
 
     public ListStorageTest() {
         super(new ListStorage());
